// src/data/generated/faceEmocoes.generated.ts
export type FaceEmocao = 'pain' | 'worried' | 'dizzy' | 'stunned' | 'sigh' | 'joyous' | 'determined'